Update release readme explanation on upmerge (#8937)
# Description

During a recent post review of the last release, one of the identified
issues was some confusion over which files should be excluded from the
upmerge process in the docs repo and the samples repo. This PR adds a
brief description to reduce confusion.

Documentation updates:

* Clarified that the PR generated by the "Upmerge docs to edge" workflow
will not include changes to `docs/config.toml` and
`docs/layouts/partials/hooks/body-end.html`, as those files are
branch-specific.
* Clarified that the PR generated by the "Upmerge samples to edge"
workflow will not include changes to `bicepconfig.json`, as that file is
branch-specific.


## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: N/A

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [x] Yes
    - [ ] Not applicable
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [x] Yes
    - [ ] Not applicable
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [x] Yes
    - [ ] Not applicable
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [x] Yes
    - [ ] Not applicable
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [x] Yes
    - [ ] Not applicable
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [x] Yes
    - [ ] Not applicable

Signed-off-by: Brooke Hamilton <45323234+brooke-hamilton@users.noreply.github.com>