Add functional test: UDT Phase 1 E2E Test (#8924)
# Description

Added e2e test for UDT (with recipe).
We define a user defined type and publish it's extension in advance. 
The test will register and deploy a bicep recipe for the userdefined
type.


## Type of change

- This pull request adds or changes features of Radius and has an
approved issue (#6688)


Fixes: Part of https://github.com/radius-project/radius/issues/6688

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [ ] Yes
    - [ ] Not applicable
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [ ] Yes
    - [ ] Not applicable
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [ ] Yes
    - [ ] Not applicable
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [ ] Yes
    - [ ] Not applicable
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [ ] Yes
    - [ ] Not applicable
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [ ] Yes
    - [ ] Not applicable

---------

Signed-off-by: lakshmimsft <ljavadekar@microsoft.com>