Add golangci-lint to dev container (#9026)
# Description

While working on [updating the go
version](https://github.com/radius-project/radius/pull/8774) there was a
need to install the correct version of `golanglint-ci`. The
[documentation](https://golangci-lint.run/welcome/install/#install-from-sources)
recommends using a binary installation, so I added that to the
devcontainer specification (`devcontainer.json`). As part of that change
I moved the `postCreateCommand` into a script. I also fixed broken links
to `golanglint-ci` in the Radius contributing docs.

Development environment setup:

*
[`.devcontainer/devcontainer.json`](diffhunk://#diff-24ad71c8613ddcf6fd23818cb3bb477a1fb6d83af4550b0bad43099813088686L45-R45):
Changed the `postCreateCommand` to use a new script
`.devcontainer/post-create.sh` instead of a long inline command.
*
[`.devcontainer/post-create.sh`](diffhunk://#diff-bcf66a848607b49d2a45edfad434f6808017ee9d18f5a20dde6670745010d89aR1-R20):
Created a new script to handle post-create tasks, including setting up
the workspace as a safe directory, installing `golangci-lint`, and other
Go tools.

Documentation updates: fix broken links

*
[`docs/contributing/contributing-code/contributing-code-first-commit/first-commit-00-prerequisites/index.md`](diffhunk://#diff-1b65ceaf649b58267d9199873c6df394b0bc30508b0bfb540cc247e0028af6a1L31-R31):
Updated the link for `golangci-lint` installation to the welcome page.
*
[`docs/contributing/contributing-code/contributing-code-first-commit/first-commit-06-creating-a-pr/index.md`](diffhunk://#diff-39e98fec72f851ba1d675c0fc5fd47452569b04c7e85331c257b790c31f967feL17-R17):
Updated the link for `golangci-lint` in the pre-PR checklist.
*
[`docs/contributing/contributing-code/contributing-code-prerequisites/README.md`](diffhunk://#diff-b3cdecf6f61c777e0a2e7089a64f8f49aeef24995ae79aa62a215f8911929cb8L102-R102):
Updated the link for `golangci-lint` in the list of core dependencies.

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: N/A

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [ ] Yes
    - [x] Not applicable
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [ ] Yes
    - [x] Not applicable
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [ ] Yes
    - [x] Not applicable
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [ ] Yes
    - [x] Not applicable

Signed-off-by: Brooke Hamilton <45323234+brooke-hamilton@users.noreply.github.com>