Make commands for workflow editing (#9642)
# Description

This PR adds a new `build/workflow.mk` file with Makefile targets for
enabling and disabling workflows, and deleting workflow runs, along with
a supporting shell script `build/workflow.sh` to implement the
operations. The `Makefile` is updated to include the new
workflow-related functionality.

These commands modify and delete data so they are hard coded to prevent
them from running on repos in the `radius-project` org.

The make commands include:

* `make workflow-disable-all`: disables all workflows in the current
repository. This is helpful when enabling GitHub actions on a fork
because it prevents scheduled workflows from running.
* `make workflow-enable-all`: opposite of `workflow-disable-all`.
* `make workflow-delete-all-runs`: deletes all workflow runs on a fork,
which is useful when testing workflow changes.

The usage output looks like this:
```shell
GitHub Workflows
  workflow-disable-all  Disable all workflows in the current repo
  workflow-enable-all   Enable all workflows in the current repo
  workflow-delete-all-runs  Delete all workflow runs in the repository. NOTE: This is a destructive operation and cannot be undone.
```

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: N/A

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [ ] Yes <!-- TaskRadio schema -->
    - [x] Not applicable <!-- TaskRadio schema -->
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [ ] Yes <!-- TaskRadio design-pr -->
    - [x] Not applicable <!-- TaskRadio design-pr -->
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [ ] Yes <!-- TaskRadio design-review -->
    - [x] Not applicable <!-- TaskRadio design-review -->
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [ ] Yes <!-- TaskRadio samples-pr -->
    - [x] Not applicable <!-- TaskRadio samples-pr -->
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [ ] Yes <!-- TaskRadio docs-pr -->
    - [x] Not applicable <!-- TaskRadio docs-pr -->
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [ ] Yes <!-- TaskRadio recipes-pr -->
    - [x] Not applicable <!-- TaskRadio recipes-pr -->

---------

Signed-off-by: Brooke Hamilton <45323234+brooke-hamilton@users.noreply.github.com>