Make command deploy AKS cluster for long-running tests (#9649)
# Description

This PR adds a new Make target that deploys the AKS cluster for
long-running tests. This allows a developer to provision a cluster that
can be used to validate changes to the LRTs. This Make target automates
the instructions at `test/infra/azure/README.md`.

* Introduced a new target `make test-deploy-lrt-cluster` to deploy an
Azure Kubernetes Service (AKS) cluster for long-running tests.
* Updated the LRT cluster deployment bicep to change the node size from
`Standard_D4as_v5` to `Standard_D8as_v5`.
* Added missing `.PHONY` to Make targets `build/test.mk`.

> NOTE: The base branch for this PR is the branch in PR #9642, so I will
rebase this branch to main when that PR is merged.

## How to Test

1. Check out the PR branch to your local dev environment.
1. Log into the Az CLI and set the desired subscription. 
1. Run `make test-deploy-lrt-cluster` to deploy an AKS cluster. A
cluster will be provisioned for you.

## Type of change

- This pull request is a minor refactor, code cleanup, test improvement,
or other maintenance task and doesn't change the functionality of Radius
(issue link optional).

Fixes: N/A

## Contributor checklist
Please verify that the PR meets the following requirements, where
applicable:

- An overview of proposed schema changes is included in a linked GitHub
issue.
    - [ ] Yes <!-- TaskRadio schema -->
    - [x] Not applicable <!-- TaskRadio schema -->
- A design document PR is created in the [design-notes
repository](https://github.com/radius-project/design-notes/), if new
APIs are being introduced.
    - [ ] Yes <!-- TaskRadio design-pr -->
    - [x] Not applicable <!-- TaskRadio design-pr -->
- The design document has been reviewed and approved by Radius
maintainers/approvers.
    - [ ] Yes <!-- TaskRadio design-review -->
    - [x] Not applicable <!-- TaskRadio design-review -->
- A PR for the [samples
repository](https://github.com/radius-project/samples) is created, if
existing samples are affected by the changes in this PR.
    - [ ] Yes <!-- TaskRadio samples-pr -->
    - [x] Not applicable <!-- TaskRadio samples-pr -->
- A PR for the [documentation
repository](https://github.com/radius-project/docs) is created, if the
changes in this PR affect the documentation or any user facing updates
are made.
    - [ ] Yes <!-- TaskRadio docs-pr -->
    - [x] Not applicable <!-- TaskRadio docs-pr -->
- A PR for the [recipes
repository](https://github.com/radius-project/recipes) is created, if
existing recipes are affected by the changes in this PR.
    - [ ] Yes <!-- TaskRadio recipes-pr -->
    - [x] Not applicable <!-- TaskRadio recipes-pr -->

---------

Signed-off-by: Brooke Hamilton <45323234+brooke-hamilton@users.noreply.github.com>
Signed-off-by: GitHub <noreply@github.com>